finished join session cell
